requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
itsdangerous
jinja2
PyJWT
passlib[bcrypt]

langchain
langchain-core
langchain-community
boto3
langchain-aws
awscli




faiss-cpu
sentence-transformers
pypdf
tqdm
pydantic

huggingface-hub

easyocr
uvicorn
passlib
pillow
react-markdown
jwt
PyJWT-2.10.1

